# api-doc/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "BossaBox App for Interview",
        "description": "BossaBox App for Interview"
    },
    "basePath": "/",
    "schemes": ["http"],
    "consumes": [
        "application/json"
      ],
      "produces": [
        "application/json"
    ],
    
    "paths": {
        "/register" : {
            "get": {
                "tags": ["Register"],
                "summary": "This endpoint doesn't exists. Should return 404 http code.",
                "responses": {
                    "404": {
                        "description": "This page doesn't exists."
                    }
                }
            },
            "post": {
                "tags": ["Register"],
                "summary": "Register new user.",
                "parameters": [
                    {
                        "name": "user",
                        "in": "body",
                        "description": "Data for new user.",
                        "schema": {
                            "$ref": "#/components/schemas/UserRegister"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "User created!",
                        "schema": {
                            "$ref": "#/components/schemas/UserLogged"
                        }
                    },
                    "200": {
                        "description": "User already exist!",
                        "schema": {
                            "$ref": "#/components/schemas/GenericMessage"
                        }
                    },
                    "500": {
                        "description": "Server error.",
                        "schema": {
                            "$ref": "#/components/schemas/GenericMessage"
                        }
                    }
                }
            }
        },
        "/login": {
            "get": {
                "tags": ["Login"],
                "summary": "This endpoint doesn't exists. Should return 404 http code.",
                "responses": {
                    "404": {
                        "description": "This page doesn't exists."
                    }
                }
            },
            "post": {
                "tags": ["Login"],
                "summary": "Register new user.",
                "parameters": [
                    {
                        "name": "user",
                        "in": "body",
                        "description": "User data for login",
                        "schema": {
                            "$ref": "#/components/schemas/UserLogin"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User login!",
                        "schema": {
                            "$ref": "#/components/schemas/UserLogged"
                        }
                    },
                    "401": {
                        "description": "User login with wrong email and password. (Unauthorized)",
                        "schema": {
                            "$ref": "#/components/schemas/GenericMessage"
                        }
                    },
                    "403": {
                        "description": "User login with wrong email and password. Maybe token is no longer valid."
                    },
                    "404": {
                        "description": "User not found.",
                        "schema": {
                            "$ref": "#/components/schemas/GenericMessage"
                        }
                    },
                    "500": {
                        "description": "Server error.",
                        "schema": {
                            "$ref": "#/components/schemas/GenericMessage"
                        }
                    }
                }
            }
        },
        "/tools" : {
            "parameters": [
                {
                    "name": "x-access-token",
                    "in": "header",
                    "required": true,
                    "description": "Access auth token",
                    "type": "string"
                }
            ],
            "get": {
                "tags": ["Tools"],
                "summary": "Get all Tools from database",
                "responses": {
                    "200": {
                        "description": "Get all Tools from database!",
                        "schema": {
                            "$ref": "#/components/schemas/Tool"
                        }
                    },
                    "401": {
                        "description": "User Unauthorized. Maybe email or password is wrong.",
                        "schema": {
                            "$ref": "#/components/schemas/GenericMessage"
                        }
                    },
                    "404": {
                        "description": "Items not found with related tag.",
                        "schema": {
                            "$ref": "#/components/schemas/GenericMessage"
                        }
                    },
                    "500": {
                        "description": "Server error.",
                        "schema": {
                            "$ref": "#/components/schemas/GenericMessage"
                        }
                    }
                }
            },
            "post": {
                "tags": ["Tools"],
                "summary": "Save new Tools",
                "parameters": [
                    {
                        "name": "user",
                        "in": "body",
                        "description": "New Tool data",
                        "schema": {
                            "$ref": "#/components/schemas/Tool"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Success!",
                        "schema": {
                            "$ref": "#/components/schemas/Tool"
                        }
                    },
                    "401": {
                        "description": "User Unauthorized. Maybe email or password is wrong.",
                        "schema": {
                            "$ref": "#/components/schemas/GenericMessage"
                        }
                    },
                    "403": {
                        "description": "Forbidden.",
                        "schema": {
                            "$ref": "#/components/schemas/JWTValidateMessage"
                        }
                    },
                    "500": {
                        "description": "Server error or Failed to authenticate token.",
                        "schema": [
                            {"$ref": "#/components/schemas/GenericMessage" },
                            {"$ref": "#/components/schemas/JWTValidateMessage"}
                        ]
                    }
                }
            }
        },
        "/tools/{id}" : {
            "parameters": [
                {
                    "name": "x-access-token",
                    "in": "header",
                    "required": true,
                    "description": "Access auth token",
                    "type": "string"
                }
            ],
            "delete": {
                "tags": ["Tools"],
                "summary": "Save new Tools",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Tool id!"
                    }
                ],
                "responses": {
                    "401": {
                        "description": "User Unauthorized. Maybe email or password is wrong.",
                        "schema": {
                            "$ref": "#/components/schemas/GenericMessage"
                        }
                    },
                    "403": {
                        "description": "Forbidden.",
                        "schema": {
                            "$ref": "#/components/schemas/JWTValidateMessage"
                        }
                    },
                    "404": {
                        "description": "Item not found in database."
                    },
                    "500": {
                        "description": "Server error while trying delete data.",
                        "schema": [
                            {"$ref": "#/components/schemas/GenericMessage" },
                            {"$ref": "#/components/schemas/JWTValidateMessage"}
                        ]
                    }
                }
            }
        },
        "/tools?tag={tag}" : {
            "parameters": [
                {
                    "name": "x-access-token",
                    "in": "header",
                    "required": true,
                    "description": "Access auth token",
                    "type": "string"
                }
            ],
            "get": {
                "tags": ["Tools"],
                "summary": "Save new Tools",
                "parameters": [
                    {
                        "name": "tag",
                        "in": "query",
                        "description": "Tag name"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tools items with related tag",
                        "schema": {
                            "$ref": "#/components/schemas/Tool"
                        }
                    },
                    "401": {
                        "description": "User Unauthorized. Maybe email or password is wrong.",
                        "schema": {
                            "$ref": "#/components/schemas/GenericMessage"
                        }
                    },
                    "403": {
                        "description": "Forbidden.",
                        "schema": {
                            "$ref": "#/components/schemas/JWTValidateMessage"
                        }
                    },
                    "404": {
                        "description": "Item not found in database."
                    },
                    "500": {
                        "description": "Server error while trying retrieve data."
                    }
                }
            }
        }
    },

    "components": {
        "schemas" : {
            "GenericMessage": {
                "properties": {
                    "msg": {
                        "type": "string"
                    }
                }
            },
            "JWTValidateMessage": {
                "properties": {
                    "msg": {
                        "type": "string"
                    },
                    "auth": {
                        "type": "boolean"
                    }
                }
            },
            "UserRegister": {
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    }
                }
            },
            "UserLogin": {
                "properties": {
                    "email": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    }
                }
            },
            "UserLogged": {
                "properties": {
                    "user": {
                        "name" : {
                            "type": "string"
                        },
                        "email": {
                            "type": "string"
                        }
                    },
                    "access_token": {
                        "type": "string"
                    },
                    "expires_in": {
                        "type": "string"
                    }
                }
            },
            "Tool": {
                "properties": {
                    "_id": "string",
                    "title": {
                        "type": "string"
                    },
                    "link": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "tags": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    }
}